Move claimToken logic
